+++ SimpleParticle.h
@@ -10,6 +10,6 @@ public:
 	//更新処理
 	void Update() override;
 	//パーティクルの追加
-	void Add(uint32_t addNum, uint32_t life, NColor color, float minScale, float maxScale,
+	void Add(uint32_t addNum, float life, NColor color, float minScale, float maxScale,
 		NVector3 minVelo, NVector3 maxVelo,NVector3 accel = {}, NVector3 minRot = {}, NVector3 maxRot = {}) override;
 };
